+++ admin/admin-server/src/main/java/com/flyemu/share/controller/MerchantController.java
@@ -47,7 +47,7 @@ public class MerchantController {
     }
 
     @DeleteMapping("/{merchantId}")
-    public JsonResult delete(@PathVariable Integer merchantId) {
+    public JsonResult delete(@PathVariable Long merchantId) {
         merchantService.delete(merchantId);
         return JsonResult.successful();
     }
