+++ admin/admin-server/src/main/java/com/flyemu/share/service/MerchantService.java
@@ -117,7 +117,7 @@ public class MerchantService extends AbsService {
 
 
     @Transactional
-    public void delete(Integer merchantId) {
+    public void delete(Long merchantId) {
         merchantRepository.deleteById(merchantId);
     }
