+++ admin/admin-server/src/main/java/com/flyemu/share/service/RoleService.java
@@ -81,7 +81,7 @@ public class RoleService extends AbsService {
      * @param roleId 角色ID
      */
     @Transactional
-    public void delete(Integer roleId, Integer merchantId) {
+    public void delete(Long roleId, Long merchantId) {
         long count = bqf.selectFrom(qAdmin).where(qAdmin.roleId.eq(roleId)).fetchCount();
         Assert.isTrue(count == 0, "已被使用不能删除~");
         jqf.delete(qMenuRole).where(qMenuRole.roleId.eq(roleId)).execute();
@@ -95,7 +95,7 @@ public class RoleService extends AbsService {
      *
      * @return
      */
-    public List<RoleSimpleDto> simpleList(Integer merchantId, Integer type) {
+    public List<RoleSimpleDto> simpleList(Long merchantId, Integer type) {
         BooleanBuilder builder = new BooleanBuilder();
         builder.and(qRole.merchantId.eq(merchantId));
         List<RoleSimpleDto> dtos = new ArrayList<>();
@@ -112,12 +112,12 @@ public class RoleService extends AbsService {
     }
 
 
-    public List<Integer> getMenuRole(Integer roleId) {
+    public List<Long> getMenuRole(Long roleId) {
         return bqf.selectFrom(qMenuRole).select(qMenuRole.menuId).where(qMenuRole.roleId.eq(roleId)).fetch();
     }
 
     @Transactional
-    public void grantMenuRole(Integer roleId, List<Integer> menus) {
+    public void grantMenuRole(Long roleId, List<Long> menus) {
         jqf.delete(qMenuRole)
                 .where(qMenuRole.roleId.eq(roleId))
                 .execute();
@@ -133,7 +133,7 @@ public class RoleService extends AbsService {
         }
     }
 
-    public List<Integer> grantMenu(Integer roleId) {
+    public List<Long> grantMenu(Long roleId) {
         return bqf.selectFrom(qMenuRole)
                 .select(qMenuRole.menuId)
                 .where(qMenuRole.roleId.eq(roleId))
@@ -144,7 +144,7 @@ public class RoleService extends AbsService {
     public static class Query {
         public final BooleanBuilder builder = new BooleanBuilder();
 
-        public void setMerchantId(Integer merchantId) {
+        public void setMerchantId(Long merchantId) {
             if (merchantId != null) {
                 builder.and(qRole.merchantId.eq(merchantId));
             }
