+++ kernel/src/main/java/com/flyemu/share/entity/MerchantMenu.java
@@ -22,13 +22,13 @@ public class MerchantMenu implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
-    private Integer menuId;
+    private Long menuId;
 
     @Column(nullable = false)
-    private Integer merchantId;
+    private Long merchantId;
 
 
 }
