+++ kernel/src/main/java/com/flyemu/share/entity/Stock.java
@@ -40,10 +40,10 @@ public class Stock {
 
     @Comment("组织ID")
     @Column(nullable = false)
-    private Integer organizationId;
+    private Long organizationId;
 
     @Comment("商户ID")
     @Column(nullable = false)
-    private Integer merchantId;
+    private Long merchantId;
 
 }
