+++ kernel/src/main/java/com/flyemu/share/repository/CustomersLevelPriceRepository.java
@@ -3,5 +3,5 @@ package com.flyemu.share.repository;
 import com.flyemu.share.entity.CustomersLevelPrice;
 import org.springframework.data.jpa.repository.support.JpaRepositoryImplementation;
 
-public interface CustomersLevelPriceRepository  extends JpaRepositoryImplementation<CustomersLevelPrice,Integer> {
+public interface CustomersLevelPriceRepository  extends JpaRepositoryImplementation<CustomersLevelPrice,Long> {
 }
