+++ kernel/src/main/java/com/flyemu/share/repository/CustomersRepository.java
@@ -12,6 +12,6 @@ import org.springframework.data.jpa.repository.support.JpaRepositoryImplementati
  * @公司信息: 纷析云（杭州）科技有限公司
  * @公司介绍: 专注于财务相关软件开发, 企业会计自动化解决方案
  */
-public interface CustomersRepository extends JpaRepositoryImplementation<Customers,Integer> {
+public interface CustomersRepository extends JpaRepositoryImplementation<Customers,Long> {
 
 }
