+++ kernel/src/main/java/com/flyemu/share/repository/MerchantUserRepository.java
@@ -10,5 +10,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @公司信息: 纷析云（杭州）科技有限公司
  * @公司介绍: 专注于财务相关软件开发, 企业会计自动化解决方案
  */
-public interface MerchantUserRepository extends JpaRepository<MerchantUser, Integer> {
+public interface MerchantUserRepository extends JpaRepository<MerchantUser, Long> {
 }
